requirements.txt
python-aiml>=0.9.3
pynacl>=1.4.0
Red-DiscordBot==3.4.0
